+++ flutter_ui_bloc/lib/src/form/fields/single/type_field_bloc_builder.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_form_bloc/flutter_form_bloc.dart';
 import 'package:flutter_ui_bloc/src/form/fields/utils/input_formatters.dart';